paraphrase-evaluation-experiment/requirements.txt
click==8.1.2
gensim==4.1.2
nltk==3.7
pyemd==0.5.1
python-dotenv==0.20.0
rouge-score==0.0.4
pandas==1.4.2
seaborn==0.11.2
tabulate==0.8.9
requests==2.27.1
statsmodels~=0.13.2
matplotlib~=3.5.1
beautifulsoup4~=4.11.1